create authentication middleware
