+++ server/DB/controllers/signin.js
@@ -1,0 +1,29 @@
+import Helper from './Helper';
+import db from '../db';
+
+const signin = {
+  async login(req, res) {
+    if (!req.body.email || !req.body.password) {
+      return res.status(400).send({ message: 'Some values are missing' });
+    }
+    if (!Helper.isValidEmail(req.body.email)) {
+      return res.status(400).send({ message: 'Please enter a valid email address' });
+    }
+    const text = 'SELECT * FROM users WHERE email = $1';
+    try {
+      const { rows } = await db.query(text, [req.body.email]);
+      if (!rows[0]) {
+        return res.status(400).send({ message: 'You are not registered yet' });
+      }
+      if (!Helper.comparePassword(rows[0].password, req.body.password)) {
+        return res.status(400).send({ message: 'Your password is incorrect' });
+      }
+      const token = Helper.generateToken(rows[0].id);
+      return res.status(200).send({ token });
+    } catch (error) {
+      return res.status(400).send(error);
+    }
+  },
+};
+
+export default signin.login;
